server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1